Factor out a Sam class that formats SAM records
Used only in align_SE for the moment.
